requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
Pillow==10.1.0
django-filter==23.3
celery==5.3.4
redis==5.0.1
django-environ==0.11.2
psycopg2-binary==2.9.9
django-crispy-forms==2.1
crispy-bootstrap5==0.7

